Piped the stats.gov query response InputStream straight to S3 PutObject
Causes a warning about unknown content length as the query response
doesn't provide it:
WARNING: No content length specified for stream data.  Stream contents
will be buffered in memory and could result in out of memory errors.

However the only way to avoid that is to buffer it in memory and measure
it in the lambda function, so no difference.